src/requirements.txt
flask
flask-cors
psycopg2
python-dotenv
SQLAlchemy
